requirements.txt
django-bmemcached==0.2.4
django-heroku==0.3.1
gunicorn==20.0.4
pylibmc==1.5.2
Django==2.2.9
django-bootstrap4==1.1.1
python-slugify==4.0.0
dj-database-url==0.5.0 
django-heroku==0.3.1 
psycopg2==2.8.4
whitenoise==5.0.1